+++ src/components/Post/style.module.css
@@ -17,6 +17,7 @@
   }
 
   span {
+    display: inline-block;
     height: var(--spacing-20);
     align-content: center;
     font-size: 0.625rem;
