+++ src/routes/home/component/MainPost.jsx
@@ -1,51 +1,57 @@
 import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import Icon from '@/components/Icon/Icon';
+// import { useNavigate } from 'react-router-dom';
+// import Icon from '@/components/Icon/Icon';
 import Post from '@/components/Post/Post';
-import PostButton from '@/components/PostButton/PostButton';
+// import PostButton from '@/components/PostButton/PostButton';
 import S from '@/routes/home/component/MainPost.module.css';
 import usePostStore from '@/stores/postStore';
 import LoadingSpinner from '../../../components/LoadingSpinner/LoadingSpinner';
+import { format, parseISO } from 'date-fns';
+import { ko } from 'date-fns/locale';
 
 function MainPost() {
-  const navigate = useNavigate();
+  // const navigate = useNavigate();
   const { posts, fetchPosts, isLoading } = usePostStore();
 
   useEffect(() => {
     fetchPosts();
   }, [fetchPosts]);
 
-  const handlePostButtonClick = () => {
-    navigate('/main/home/new/post');
-  };
+  // const handlePostButtonClick = () => {
+  //   navigate('/main/home/new/post');
+  // };
 
   if (isLoading) return <LoadingSpinner />;
 
   return (
     <div className={S.component}>
-      {posts.map((post) => (
-        <Post
-          key={post.id}
-          title={post.title}
-          date={post.date}
-          time={post.time}
-          place={post.location}
-          member={Number(post.memberCount) || 0}
-          category={post.category}
-          id={post.id}
-          writer={post.expand?.writer?.name || 'Unknown'}
-        />
-      ))}
+      {posts.map((post) => {
+        let formattedDate = '날짜 없음';
+        try {
+          if (post.date) {
+            const date =
+              typeof post.date === 'string'
+                ? parseISO(post.date)
+                : new Date(post.date);
+            formattedDate = format(date, 'yyyy.MM.dd. a HH:mm', { locale: ko });
+          }
+        } catch (error) {
+          console.error('날짜 포맷팅 오류:', error);
+        }
 
-      <PostButton
-        zIndex="1000"
-        bottom="15.3rem"
-        right="15.3rem"
-        borderRadius="3.125rem"
-        onClick={handlePostButtonClick} // 클릭 시 이동
-      >
-        <Icon id="calendarPlus" color={'var(--white)'}></Icon>
-      </PostButton>
+        return (
+          <Post
+            key={post.id}
+            title={post.title}
+            date={formattedDate}
+            place={post.location}
+            member={Number(post.memberCount) || 0}
+            category={post.category}
+            id={post.id}
+            writer={post.expand?.writer?.name || 'Unknown'}
+          />
+        );
+      })}
     </div>
   );
 }
